feat: validation and return message updated

+++ src/modules/user/useCase/updateUserAvatar/updateUserAvatarUseCase.ts
@@ -1,3 +1,5 @@
+import { IUsersTokensRepository } from 'modules/user/infra/repositories/IUsersTokensRepository';
+import { AppError } from '../../../../errors/appErrors';
 import { deleteFile } from '../../../../utils/file';
 
 import { IUsersRepository } from '../../infra/repositories/IUsersRepository';
@@ -12,6 +14,15 @@ export class UpdateUserAvatarUseCase {
 
   async execute({ user_id, avatarFile }: IRequest): Promise<void> {
     const user = await this.usersRepository.findById(user_id);
+    if (!user) {
+      throw new AppError("User not found");
+    }
+   
+    // const userToken = await this.usersTokensRepository.findByUserTokenId(user_id);
+    // if (!userToken) {
+    //   throw new AppError("User not found");
+    // }
+
     if (user.avatar) {
       await deleteFile(`./tmp/avatar/${user.avatar}`);
     }
